+++ flutter/storyly_flutter/example/lib/main.dart
@@ -36,10 +36,11 @@ class _MyAppState extends State<MyApp> {
             androidParam: StorylyParam()
             ..storylyId = YOUR_APP_ID_FROM_DASHBOARD,
             iosParam: StorylyParam()
-            ..storylyId = YOUR_APP_ID_FROM_DASHBOARD,
+            ..storylyId = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJhY2NfaWQiOjg0NiwiYXBwX2lkIjo0NzAsImluc19pZCI6NDY2fQ.5Ag5dqOxAWzKwYnuM1Voo_bTl7RMXFRh2S1EzVeI01w",
             storylyLoaded: (storyGroupList) => print("storylyLoaded"),
             storylyLoadFailed: (errorMessage) => print("storylyLoadFailed"),
             storylyActionClicked: (story) => print("storylyActionClicked"),
+            storylyEvent: (event) => print("storylyEvent"),
             storylyStoryShown: () => print("storylyStoryShown"),
             storylyStoryDismissed: () => print("storylyStoryDismissed"),
             storylyUserInteracted: (eventPayload) => print("storylyUserInteracted")